created features and shared directories and moved files into features
